User can now click and drag nodes on the map.
Some functionality had to be re-addressed here.
Behaviour is changed so that the node moves independently until the mouseup event is fired.

With some retouches to ensure that react is providing the right state, trying to do this in the mousemove means that the whole map is re-rendered and we lose the events.

This will be worth re-visiting once everthing is living inside its own component.