Fixed issues
The initial issue (map vanishing on drag) was actually a revisit on a previous issue where mapText was updated after it was called for. - Fixed by putting calls back into effects.
This caused a knockon with effects not behaving as expected. - I've straightened these out, however there is plenty of opporunity for refactor here.